feat: update coin balances in higher level
